pyproject.toml
[project]
name = "text2sql"
version = "0.1.0"
description = "A tool to convert natural language queries into SQL statements."
authors = [
    {name = "niraj.kumbhar"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "unstructured[md] (>=0.17.2,<0.18.0)",
    "langchain-openai (>=0.3.18,<0.4.0)",
    "mysql-connector-python (>=9.4.0,<10.0.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "streamlit (>=1.45.1,<2.0.0)",
]

[tool.poetry]
package-mode = false


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

